add snmpv3
To create a v3 client, a lot more things need to be configured,
and the caller needs to know to set an appropriate context, rather
than using special community strings as in v2c.  snmerp doesn't pretend
to be an abstraction layer here.

Currently implements aes-128 for privacy and sha-1 for authentication.
aes-256 and sha-2 may follow if someone makes me implement it.

For now, this uses the snmp library to convert passwords to localized keys.
If I need to add algorithms that snmp:passwd2localized_key doesn't support,
that will change.